Remove RESET_LOG; Add check for dataLength log entry return to CRReplication

+++ src/common_info.h
@@ -41,7 +41,7 @@ enum MessageType {
 };
 
 struct LogEntryHeader {
-    size_t popcnt;
+    uint64_t popcnt;
     LogEntryState state{DIRTY};
     uint64_t dataLength;
 };
@@ -54,7 +54,7 @@ enum MessageType {
     TERMINATE
 };
 struct LogEntryHeader {
-    size_t popcnt;
+    uint64_t popcnt;
     uint64_t dataLength;
 };
 #endif
